+++ nengo/tests/test_learning_rules.py
@@ -3,7 +3,8 @@ import pytest
 
 import nengo
 from nengo.dists import UniformHypersphere
-from nengo.learning_rules import LearningRuleTypeParam, PES, BCM, Oja, Voja
+from nengo.learning_rules import (LearningRuleTypeParam, PES, BCM, Oja, Voja,
+                                  GenericRule)
 from nengo.processes import WhiteSignal
 
 
@@ -466,3 +467,113 @@ def test_frozen():
 
     with pytest.raises((ValueError, RuntimeError)):
         a.learning_rate = 1e-1
+
+
+@pytest.mark.parametrize('modifies', ["decoders", "weights"])
+def test_generic_rule(Simulator, nl_nodirect, seed, modifies):
+    n = 100
+    input_vector = np.asarray([0.5, -0.5])
+    dt = 0.001
+    reg = 0.1
+
+    m = nengo.Network(seed=seed)
+    with m:
+        m.config[nengo.Ensemble].neuron_type = nl_nodirect()
+        u = nengo.Node(output=input_vector)
+        a = nengo.Ensemble(n, dimensions=2)
+        b = nengo.Ensemble(n, dimensions=2)
+
+        nengo.Connection(u, a)
+
+        if modifies == "decoders":
+            def regularized_pes(target, data):
+                return -dt / n * np.outer(data[:2], data[2:]) - target * reg
+        else:
+            def regularized_pes(target, data, params):
+                encoders = params[b].encoders
+                gains = params[b].gain
+                return (-dt / n *
+                        np.outer(np.dot(encoders, data[:2]) * gains,
+                                 data[2:]) - target * reg)
+
+        rule = GenericRule(regularized_pes, learning_rate=1e-3,
+                           size_in=2 + n, modifies=modifies,
+                           pass_model_params=modifies == "weights")
+
+        conn = nengo.Connection(
+            a if modifies == "decoders" else a.neurons,
+            b if modifies == "decoders" else b.neurons,
+            transform=1 if modifies == "decoders" else
+            nengo.dists.Gaussian(0, 0.001),
+            learning_rule_type=rule)
+        nengo.Connection(b, conn.learning_rule[:2])
+        nengo.Connection(u, conn.learning_rule[:2])
+        nengo.Connection(a.neurons, conn.learning_rule[2:], synapse=0.005)
+
+        b_p = nengo.Probe(b, synapse=0.1)
+        weights_p = nengo.Probe(conn, attr="weights")
+
+    with Simulator(m, dt=dt) as sim:
+        sim.run(1.)
+
+    tmask = sim.trange() > .9
+    assert (np.sum(np.abs(sim.data[weights_p][:10])) >
+            np.sum(np.abs(sim.data[weights_p][-10:])))
+    assert np.allclose(sim.data[b_p][tmask], -input_vector, atol=0.05)
+
+
+def test_generic_encoders(Simulator, nl_nodirect, rng, seed):
+    """Voja test implemented via generic rule."""
+
+    n = 200
+    learned_vector = np.asarray([0.3, -0.4, 0.6])
+    learned_vector /= np.linalg.norm(learned_vector)
+    n_change = n // 2  # modify first half of the encoders
+    dt = 1e-3
+
+    # Set the first half to always fire with random encoders, and the
+    # remainder to never fire due to their encoder's dot product with the input
+    intercepts = np.asarray([-1] * n_change + [0.99] * (n - n_change))
+    rand_encoders = UniformHypersphere(surface=True).sample(
+        n_change, len(learned_vector), rng=rng)
+    encoders = np.append(
+        rand_encoders, [-learned_vector] * (n - n_change), axis=0)
+
+    m = nengo.Network(seed=seed)
+    with m:
+        m.config[nengo.Ensemble].neuron_type = nl_nodirect()
+        u = nengo.Node(output=learned_vector)
+        x = nengo.Ensemble(n, dimensions=len(learned_vector),
+                           intercepts=intercepts, encoders=encoders,
+                           radius=2.0)  # to test encoder scaling
+
+        def generic_voja(enc, data, params):
+            scale = (params[x].gain / x.radius)[:, None]
+            return (scale * np.outer(data[:n], data[n:]) -
+                    data[:n, None] * enc)
+
+        conn = nengo.Connection(
+            u, x, synapse=None,
+            learning_rule_type=GenericRule(
+                generic_voja, learning_rate=1e-1 * dt, modifies="encoders",
+                size_in=len(learned_vector) + n, pass_model_params=True))
+        nengo.Connection(u, conn.learning_rule[n:], synapse=None)
+        nengo.Connection(x.neurons, conn.learning_rule[:n], synapse=0.005)
+
+    with Simulator(m, dt=dt) as sim:
+        sim.run(1.0)
+
+        scaled_encoders = sim.signals[sim.model.sig[x]['encoders']]
+
+    encoder_scale = (sim.model.params[x].gain / x.radius)[:, np.newaxis]
+
+    # Check that the last half kept the same encoders throughout the simulation
+    assert np.allclose(sim.model.params[x].scaled_encoders[n_change:],
+                       scaled_encoders[n_change:])
+    # and that they are also equal to their originally assigned value
+    assert np.allclose(sim.model.params[x].scaled_encoders[n_change:] /
+                       encoder_scale[n_change:], -learned_vector)
+
+    # Check that the first half converged to the input
+    assert np.allclose(scaled_encoders[:n_change] / encoder_scale[:n_change],
+                       learned_vector, atol=0.01)