+++ CHANGES.rst
@@ -47,6 +47,8 @@ Release History
   `#1077 <https://github.com/nengo/nengo/pull/1077>`_)
 - Tweaked ``rasterplot`` so that spikes from different neurons don't overlap.
   (`#1121 <https://github.com/nengo/nengo/pull/1121>`_)
+- Added ``GenericRule`` to allow for easy implementation of new learning rules.
+  (`#553 <https://github.com/nengo/nengo/pull/553>`_)
 
 **Bug fixes**